third_party/WebKit/LayoutTests/bluetooth/descriptor/writeValue/gen-io-op-device-disconnects-during-error.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(
    () => {
      let val = new Uint8Array([1]);
      return setBluetoothFakeAdapter('FailingGATTOperationsAdapter')
          .then(() => requestDeviceWithKeyDown({
                  filters: [{services: [errorUUID(0xA0)]}],
                  optionalServices: [request_disconnection_service_uuid]
                }))
          .then(device => device.gatt.connect())
          .then(gattServer => {
            let error_characteristic;
            return gattServer.getPrimaryService(errorUUID(0xA0))
                .then(es => es.getCharacteristic(errorUUID(0xA1)))
                .then(
                    characteristic => characteristic.getDescriptor(
                        'gatt.characteristic_user_description'))
                .then(d => user_description = d)
                .then(() => get_request_disconnection(gattServer))
                .then(requestDisconnection => {
                  requestDisconnection();
                  return assert_promise_rejects_with_message(
                      user_description.writeValue(val),
                      new DOMException(
                          'GATT Server disconnected while performing a GATT operation.',
                          'NetworkError'));
                });
          });
    },
    'Device disconnects during a writeValue call that fails. ' +
        'Reject with NetworkError.');

</script>
